% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data based on span}
\usage{
aggregate_data(data_trimmed, span)
}
\arguments{
\item{data_trimmed:}{A dataframe of time-series data with columns named Date, Time, Measurement_1, Measurement_2, and so on.}

\item{span:}{A numeric value that represents the time span in minutes. It specifies the time interval for grouping the data.}
}
\value{
The function converts the span value to minutes and determines the grouping for the data based on the time span. If the time span is less than or equal to 30 minutes, no aggregation is performed, and the original data is returned. If the time span is greater than 30 minutes, the data is aggregated based on the specified time span.

The function uses the aggregate function to group the data and calculate the mean of each group for the data columns. The function then converts the Date column of the aggregated dataframe to the appropriate date format based on the grouping interval.

The function returns an aggregated dataframe with the same columns as the input dataframe, but with the data aggregated based on the specified time span.
}
\description{
aggregate_data is a function that aggregates time-series data based on a specified time span. It takes a dataframe of time-series data, and a time span (in minutes) as arguments and returns an aggregated dataframe. The function groups the data based on the specified time span and calculates the mean of each group for the data columns.
}
