% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_graphs.R
\name{as_alert_graphs}
\alias{as_alert_graphs}
\title{AS Alert Graphs}
\usage{
as_alert_graphs(alertsDF, data, folder_name)
}
\arguments{
\item{alertsDF}{Dataframe containing 1 row per alert with the client name (or client campaign combo) and the alerts column.}

\item{data}{The dataframe containing the daily data used to run the model and check for alerts.}

\item{folder_name}{Folder name of where to save the plots}
}
\value{
Graphs for all non-"Negative RexT" alerts using the given model.
}
\description{
Generate graphs for all alerts using a given model for visual inspection.
}
\author{
Stefanie Molin
}
