% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_graphs.R
\name{as_metric_graphs}
\alias{as_metric_graphs}
\title{AS Metric Graphs}
\usage{
as_metric_graphs(combosDF, data, check_for_alert_date, folder_name)
}
\arguments{
\item{combosDF}{Dataframe containing per client name (or client campaign combo).}

\item{data}{The dataframe containing the daily data used to run the model and check for alerts.}

\item{check_for_alert_date}{The date you went to check for anomalies.}

\item{folder_name}{Folder name of where to save the plots}
}
\value{
Graphs for all possible combinations of client (or campaign) and metric saved in specified folder
and returns all the client + metric combos and whether or not eligible they are for detection that day.
}
\description{
Generate graphs for all alerts using a given model for visual inspection.
}
\author{
Stefanie Molin
}
