% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{AS_transparent_graph}
\alias{AS_transparent_graph}
\title{AS Transparent Graph}
\usage{
AS_transparent_graph(data, metric, client_name, encode = FALSE,
  max_possible_days = 60)
}
\arguments{
\item{data}{Dataframe containing daily data.}

\item{metric}{The name of the metric to graph, (must match column names, but not case-sensitive).}

\item{client_name}{Name of the client for which the graph is being created.}

\item{encode}{Whether or not to base64 encode the graph, defaults to FALSE. 
(Needed if sending as part of HTML attachment rather than email body inline images.)}

\item{max_possible_days}{Controls the shifting of the label by creating a ratio of max days allowed in the graph (defaults to 60) and the total days being graphed. 
This makes sure that graphs with less than the max days will shift the label less to keep it on the line.}
}
\value{
A base64 encoded site events plot, or just the plot object (depends on the value of encode).
}
\description{
Make a ggplot2 graph for given metric in AS alerts
}
\author{
Stefanie Molin
}
