% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{base64_encode_graph}
\alias{base64_encode_graph}
\title{Base64 Encode a Graph}
\usage{
base64_encode_graph(plot_object)
}
\arguments{
\item{plot_object}{Plot to encode.}
}
\value{
base64 encoding of the plot
}
\description{
Use base64 to encode an image, then remove the <img> tags for easy use in HTML.
}
\author{
Stefanie Molin
}
