% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{build_and_send_AS_alerts}
\alias{build_and_send_AS_alerts}
\title{Build and Send AS Alerts}
\usage{
build_and_send_AS_alerts(client_stats, client_alerts, root_config_list,
  regional_config_list, run_date, test_mode, sender, mailbox_password)
}
\arguments{
\item{client_stats}{Dataframe containing all the data collected during the run}

\item{client_alerts}{Dataframe containing all the alerts triggered for the day}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{run_date}{Date for which the alerts were calculated (today's date in most cases)}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}
}
\value{
Builds and sends SCOPE alerts for all AS with alerts on their accounts. No return.
}
\description{
Build and send all SCOPE AS alerts
}
\author{
Stefanie Molin
}
