% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{build_and_send_exec}
\alias{build_and_send_exec}
\title{Build and Send Exec}
\usage{
build_and_send_exec(alerts, email_data, accounts_list, plot_list,
  highlight_color, regular_color, regional_config_list, root_config_list,
  template, test_mode, sender, mailbox_password)
}
\arguments{
\item{alerts}{String of alerts}

\item{email_data}{Dataframe of email data}

\item{accounts_list}{List containing accounts to use to determine gainers/losers section.}

\item{plot_list}{Nested lists containing file_name and plot for each region}

\item{highlight_color}{String containing the hexcode of the highlighted text color}

\item{regular_color}{String containing the hexcode of the regular text color}

\item{regional_config_list}{List of regional_config settings.}

\item{root_config_list}{List from the root_config file.}

\item{template}{Path to the HTML template.}

\item{test_mode}{Boolean indicating whether or not SCOPE Exec is running in test mode.}

\item{sender}{Email address to send from}

\item{mailbox_password}{Password for the sender}
}
\description{
Build and send the SCOPE Exec email.
}
\author{
Stefanie Molin
}
