% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{build_scope_exec}
\alias{build_scope_exec}
\title{Build SCOPE Exec}
\usage{
build_scope_exec(alerts, email_data, accounts_list, plot_list, highlight_color,
  regular_color, regional_config_list, template)
}
\arguments{
\item{alerts}{String of alerts}

\item{email_data}{Results data for email}

\item{accounts_list}{List containing accounts to use to determine gainers/losers section.}

\item{plot_list}{Nested lists containing file_name and plot for each region}

\item{highlight_color}{String containing the hexcode of the highlighted text color}

\item{regular_color}{String containing the hexcode of the regular text color}

\item{regional_config_list}{List of regional_config settings.}

\item{template}{HTML template to replace values in}
}
\value{
HTML
}
\description{
Fill in the Exec template
}
\author{
Stefanie Molin
}
