% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ET.R
\name{calculate_ET}
\alias{calculate_ET}
\title{Calculate evapotranspiration (ET) from latent heat of the canopy (lEctot) and average canopy temperature (tcave)}
\usage{
calculate_ET(lEctot, tcave)
}
\arguments{
\item{lEctot}{Latent heat of the canopy in MJ/m^2}

\item{tcave}{Average canopy temperature in Celsius}
}
\value{
Evapotranspiration (ET) in meters
}
\description{
Calculate evapotranspiration (ET) from latent heat of the canopy (lEctot) and average canopy temperature (tcave)
}
\details{
This function calculates evapotranspiration (ET) using the Penman-Monteith equation, which takes into account the latent heat of the canopy and average canopy temperature.
The function converts the latent heat from MJ/m^2 to kg/m^2 and calculates ET based on the formula ET = lEctot_kg / (1000 * (tcave + 273.15)),
where lEctot_kg is the latent heat in kilograms per square meter and tcave is the average canopy temperature in Celsius.
The result is returned as evapotranspiration (ET) in meters.
}
\examples{
lEctot <- 1500  # Latent heat of the canopy in MJ/m^2
tcave <- 25     # Average canopy temperature in Celsius
ET <- calculate_ET(lEctot, tcave)
print(ET)

}
