% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_output.R
\name{call_output}
\alias{call_output}
\title{This calls out the filepath of the type of SCOPE output you need for further analysis}
\usage{
call_output(type)
}
\description{
This calls out the filepath of the type of SCOPE output you need for further analysis
}
