% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{capitalize_name}
\alias{capitalize_name}
\title{Capitalize Name}
\usage{
capitalize_name(name)
}
\arguments{
\item{name}{String to modify}
}
\value{
String
}
\description{
Capitalize the first letter of each word in a string
}
\author{
Stefanie Molin
}
