% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{change_in_bps}
\alias{change_in_bps}
\title{Basis Points Change}
\usage{
change_in_bps(before_values, after_values, multiplier = 1, format = TRUE,
  ...)
}
\arguments{
\item{before_values}{Vector of values in "before" condition}

\item{after_values}{Vector of values in "after" condition}

\item{multiplier}{Value to multiple the difference by (need to use if both the input values were modified elsewhere)}

\item{format}{Whether or not to format as bps}

\item{...}{Additional arguments to be passed to format_bps()}
}
\value{
Vector of characters
}
\description{
Calculate basis point change and return formatted if desired
}
\author{
Stefanie Molin
}
