% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{check_history}
\alias{check_history}
\title{Check If Script Ran}
\usage{
check_history(history_file, script_name, geo = "")
}
\arguments{
\item{history_file}{Path of the YAML file containing the run history}

\item{script_name}{Name of the script to check for}

\item{geo}{Location to check for}
}
\value{
Boolean indicating whether or not the script_name / geo combination has already run
}
\description{
Check which scripts have run in which geos
}
\author{
Francois Pillot

Stefanie Molin
}
