% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomaly.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(df)
}
\arguments{
\item{df}{Dataframe to clean}
}
\value{
A dataframe
}
\description{
Replace Inf, NaN and NA values with 0
}
\author{
Stefanie Molin
}
