% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{clean_omics_data}
\alias{clean_omics_data}
\title{Clean Omics Data}
\usage{
clean_omics_data(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Cleaned data frame
}
\description{
Basic function to impute missing values and scale data.
}
