% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{class}
\name{ClonalAnalysis-class}
\alias{ClonalAnalysis-class}
\alias{ClonalAnalysis}
\title{Output of the analyzeClones function}
\description{
\code{ClonalAnalysis} contains output from the \link{analyzeClones} function.
It includes infromation to interpret clonal assignment performance.
}
\section{Slots}{

\describe{
\item{\code{threshold}}{cut-off separating the inter (within) and intra (between)
clonal distances.}

\item{\code{inter_intra}}{data.frame containing all inter and intra clonal distances.}

\item{\code{plot_inter_intra}}{density plot of inter versus intra clonal distances. The threshold is
shown with a horizental dashed-line.}

\item{\code{neighborhoods}}{a numeric vector containing scale parameters used in spectral
clustering process.}

\item{\code{plot_neighborhoods}}{histogram of neighborhoods. The threshold is shown with a vertical
dashed-line.}
}}

\seealso{
\link{analyzeClones}
}
