% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_convert_photo_to_gpp.R
\name{convert_photo_to_gpp}
\alias{convert_photo_to_gpp}
\title{Function to convert umol CO2 m-2 s-1 to g C halfhour-1}
\usage{
convert_photo_to_gpp(actot, area_m2)
}
\arguments{
\item{actot}{canopy photosynthesis}

\item{area_m2}{area of a leaf (1m2 is default)}
}
\description{
Function to convert umol CO2 m-2 s-1 to g C halfhour-1
}
