% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_convert_photo_to_gpp_day.R
\name{convert_photo_to_gpp_day}
\alias{convert_photo_to_gpp_day}
\title{Function to convert umol CO2 m-2 s-1 to g C day}
\usage{
convert_photo_to_gpp_day(actot, area_m2)
}
\arguments{
\item{actot}{Canopy photosynthesis}

\item{area_m2}{area of a leaf}
}
\description{
Function to convert umol CO2 m-2 s-1 to g C day
}
