% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_convert_to_PAR.R
\name{convert_to_PAR}
\alias{convert_to_PAR}
\title{convert solar radiation to PAR}
\usage{
convert_to_PAR(solar_radiation)
}
\arguments{
\item{solar_radiation}{solar radiation in Wm-2}
}
\description{
convert solar radiation to PAR
}
