% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Creates input for time series}
\usage{
create_matrix(par_variable, par_constant)
}
\arguments{
\item{par_variable}{a dataframe containing environmental variables}

\item{par_constant}{a dataframe containing leaf and canopy traits that are constant}
}
\description{
This function uses start and end times, a grain of resolution, and constant
and variable parameter values to create the time series input file for SCOPE
time series function. NOTE: Berkeley time format is YYYYMMDDHHMM but this
function only takes input as "\%Y-\%m-\%d \%H:\%M"
}
