% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_check_library.R
\name{data_late}
\alias{data_late}
\title{Check if Data is Late}
\usage{
data_late(data, table_name)
}
\arguments{
\item{data}{Dataframe containing 2 columns: the name of the table (table_name) and the update_time. This table is the result of the check dates query.}

\item{table_name}{Name(s) of the table to check (must be in data)}
}
\value{
Boolean indicating whether or not that table is late.
}
\description{
Checks the results of the check dates query for a specific table and determines if that table is late
}
\author{
Stefanie Molin
}
