% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt File}
\usage{
decrypt(filename)
}
\arguments{
\item{filename}{Path to the file to decrypt}
}
\value{
Dataframe with the results
}
\description{
Decrypts a file that was encrypted using a common key.
}
\author{
Stefanie Molin
}
