% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_doy_to_date.R
\name{doy_to_date}
\alias{doy_to_date}
\title{Function to Convert DoY and Year to Julian Date}
\usage{
doy_to_date(doy, year)
}
\arguments{
\item{doy}{day of year with fraction}
}
\description{
Function to Convert DoY and Year to Julian Date
}
