% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_doy_to_fract.R
\name{doy_to_fract}
\alias{doy_to_fract}
\title{converts doy to fraction of a day}
\usage{
doy_to_fract(doy)
}
\arguments{
\item{doy}{Day of year with fraction}
}
\description{
converts doy to fraction of a day
}
