% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_doy_to_time.R
\name{doy_to_time}
\alias{doy_to_time}
\title{Function to Convert DoY to time of day}
\usage{
doy_to_time(doy)
}
\arguments{
\item{doy}{day of year fraction}
}
\description{
Function to Convert DoY to time of day
}
