% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt File}
\usage{
encrypt(df, filename)
}
\arguments{
\item{df}{Dataframe containing the login and password in that order.}

\item{filename}{Where to save the encrypted credentials}
}
\value{
Dataframe with the results
}
\description{
Encrypts a file using a common key.
}
\author{
Stefanie Molin
}
