% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_account_df_builder}
\alias{exec_account_df_builder}
\title{Exec Account Dataframe Builder}
\usage{
exec_account_df_builder(df, isGainer)
}
\arguments{
\item{df}{Dataframe with gainer or loser data}

\item{isGainer}{Boolean indicating whether data is for gainers or not.}
}
\value{
Dataframe
}
\description{
Creates the gainers or losers dataframe for Exec
}
\author{
Stefanie Molin
}
