% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_account_df_empty_row}
\alias{exec_account_df_empty_row}
\title{Generate Exec Client Empty Row}
\usage{
exec_account_df_empty_row()
}
\value{
1-row dataframe for account RexT alerts section of Exec
}
\description{
Initializes a 1-row dataframe for holding account RexT alerts for Exec
}
\author{
Stefanie Molin
}
