% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_account_df_filler}
\alias{exec_account_df_filler}
\title{Exec Account Dataframe Filler}
\usage{
exec_account_df_filler(df)
}
\arguments{
\item{df}{Dataframe with Exec account data}
}
\value{
Dataframe with 5 rows
}
\description{
Fills in empty rows for any Exec account dataframes with less than 5 rows
}
\author{
Stefanie Molin
}
