% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_df_binder}
\alias{exec_df_binder}
\title{Create Exec Territory Dataframe}
\usage{
exec_df_binder(dates_vector, rext_list)
}
\arguments{
\item{dates_vector}{Vector of the dates to use for all series}

\item{rext_list}{A list of RexT series where the name is the series name how it should appear in the email and the value is the corresponding RexT vector.}
}
\value{
Dataframe with the results of binding together information
}
\description{
Turn vectors of stats by region into a dataframe
}
\author{
Stefanie Molin
}
