% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{exec_fill_in_plots}
\alias{exec_fill_in_plots}
\title{Exec Fill in Plots}
\usage{
exec_fill_in_plots(plot_list, templateHTML)
}
\arguments{
\item{plot_list}{Nested lists containing file_name and plot for each region}

\item{templateHTML}{HTML template to replace values in}
}
\value{
HTML
}
\description{
Fill the plots into the template.
}
\author{
Stefanie Molin
}
