% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_find_gainers_losers}
\alias{exec_find_gainers_losers}
\title{Find Exec Client Gainers and Losers}
\usage{
exec_find_gainers_losers(df)
}
\arguments{
\item{df}{Dataframe with client RexT data}
}
\value{
List containing 2 dataframes: gainers and losers
}
\description{
Identifies gainers and losers
}
\author{
Stefanie Molin
}
