% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{exec_graph}
\alias{exec_graph}
\title{Exec Territory Graphs}
\usage{
exec_graph(df, nudge_list = list(series = "LATAM", isSeriesNudge = 20000,
  isNotSeriesNudge = -5000), label_date_range)
}
\arguments{
\item{df}{dataframe for the exec graph}

\item{nudge_list}{Series, isSeriesNudge, and isNotSeriesNudge to pass to ggrepel for labeling}

\item{label_date_range}{vector of dates to show labels for}
}
\value{
A ggplot
}
\description{
Make a ggplot2 graph for RexT series using ggrepel to label
}
\examples{
\dontrun{
americasPlot <- exec_graph(df = americasPlotDF,
                           nudge_list = list(Series = "LATAM", isSeriesNudge = 20000, isNotSeriesNudge = -5000),
                           label_date_range = c(yesterday, dayBeforeYesterday))
}

}
\author{
Stefanie Molin
}
