% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{exec_highlight_text}
\alias{exec_highlight_text}
\title{Exec Highlight Text}
\usage{
exec_highlight_text(colors_list, templateHTML)
}
\arguments{
\item{colors_list}{List containing the hexcode for each region's RexT}

\item{templateHTML}{HTML template to replace values in}
}
\value{
HTML
}
\description{
Fill in the text color for the RexT data by region.
}
\author{
Stefanie Molin
}
