% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_initialize_account_df}
\alias{exec_initialize_account_df}
\title{Initialize Exec Account Dataframe}
\usage{
exec_initialize_account_df(dataIsLate = FALSE)
}
\arguments{
\item{dataIsLate}{Boolean indicating whether or not account data is late. Defaults to FALSE}
}
\value{
Dataframe for account RexT alerts section of Exec
}
\description{
Initializes a dataframe for holding account RexT alerts for Exec
}
\author{
Stefanie Molin
}
