% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_late_data_account_list}
\alias{exec_late_data_account_list}
\title{Create Exec Late Data Account List}
\usage{
exec_late_data_account_list()
}
\value{
List containing 2 dataframes: gainers and losers
}
\description{
Creates account lists for when data is late
}
\author{
Stefanie Molin
}
