% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{exec_series_shrink_round}
\alias{exec_series_shrink_round}
\title{Round and Shrink Exec Data Lists}
\usage{
exec_series_shrink_round(rext_list, last_x_days, decimal_places)
}
\arguments{
\item{rext_list}{List of RexT vectors}

\item{last_x_days}{Last x days to isolate in the vectors}

\item{decimal_places}{The number of decimal places to round the RexT numbers.}
}
\value{
List where all entries have been rounded and number of entries reduced to last x
}
\description{
For each vector in the rext_list reduce it to the last x days and round all entries to specified decimal places
}
\author{
Stefanie Molin
}
