% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{exec_text_color}
\alias{exec_text_color}
\title{Exec Text Coloring}
\usage{
exec_text_color(alerted_regions, highlight_color, regular_color)
}
\arguments{
\item{alerted_regions}{String containing the alerts triggered}

\item{highlight_color}{String containing the hexcode of the highlighted text color}

\item{regular_color}{String containing the hexcode of the regular text color}
}
\value{
HTML
}
\description{
Determine which color to show each regions numbers in.
}
\author{
Stefanie Molin
}
