% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{fill_in_exec_account_section}
\alias{fill_in_exec_account_section}
\title{Fill in Exec Account Section}
\usage{
fill_in_exec_account_section(accounts_list, region, templateHTML)
}
\arguments{
\item{accounts_list}{List containing the dataframes for gainers and losers}

\item{region}{String of the region being supplied}

\item{templateHTML}{HTML template to replace values in}
}
\value{
HTML
}
\description{
Fill in the accounts section of the Exec template for a given region.
}
\author{
Stefanie Molin
}
