% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating_library.R
\name{fill_in_exec_regions}
\alias{fill_in_exec_regions}
\title{Fill in Exec Regions}
\usage{
fill_in_exec_regions(data, templateHTML)
}
\arguments{
\item{data}{Dataframe of results}

\item{templateHTML}{HTML template to replace values in}
}
\value{
HTML
}
\description{
Fill in the region section of the Exec template.
}
\author{
Stefanie Molin
}
