% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomaly.R
\name{fill_in_missing_dates}
\alias{fill_in_missing_dates}
\title{Fill In Missing Dates}
\usage{
fill_in_missing_dates(data, series_key_column, metric_columns)
}
\arguments{
\item{data}{Dataframe to check and fill for missing dates (see note for structure)}

\item{series_key_column}{The name of the column that defines a unique time series (to separate from the rest in the dataframe)}

\item{metric_columns}{The names of the columns with metrics to impute}
}
\value{
A dataframe
}
\description{
For any missing dates between the minimum and maximum dates in the time series, impute the median for the metrics.
}
\note{
There must be a column "day" in data of type "Date"
}
\author{
Stefanie Molin
}
