% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metis_data.R
\name{filter_metis_data}
\alias{filter_metis_data}
\title{Filter data for Metis graphs}
\usage{
filter_metis_data(filename, rows_threshold = 3)
}
\arguments{
\item{filename}{Path to the file to use}

\item{rows_threshold}{Number of days series must have to stay in data}
}
\description{
Filter a file of Metis data
}
\author{
Stefanie Molin
}
