% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{format_bps}
\alias{format_bps}
\title{Format in Basis Points (bps)}
\usage{
format_bps(values, showPlusSign = FALSE, digits = 0)
}
\arguments{
\item{values}{Vector of values to be formatted (as basis points)}

\item{showPlusSign}{Whether or not to show "+" in front of positive percentages}

\item{digits}{How many digits to round to and display}
}
\value{
Vector of characters
}
\description{
Format values with "bps" at end, x digits after the decimal and comma separators at thousands, optional "+" in front
}
\author{
Stefanie Molin
}
