% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_format_datetime.R
\name{format_datetime}
\alias{format_datetime}
\title{Function to merge date and time}
\usage{
format_datetime(
  date,
  time,
  format = "\%Y-\%m-\%d \%H:\%M",
  tz = "America/Panama"
)
}
\arguments{
\item{date}{Date}

\item{time}{Time}

\item{format}{format}

\item{tz}{timezone}
}
\description{
Function to merge date and time
}
