% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{format_number}
\alias{format_number}
\title{Format Numbers}
\usage{
format_number(values, digits = 2)
}
\arguments{
\item{values}{Vector of values to be formatted}

\item{digits}{How many digits to round to and display}
}
\value{
Vector of characters
}
\description{
Format values with x digits after the decimal and comma separators at thousands
}
\author{
Stefanie Molin
}
