% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{format_percent}
\alias{format_percent}
\title{Format as Percent}
\usage{
format_percent(values, showPlusSign = FALSE, digits = 2)
}
\arguments{
\item{values}{Vector of values to be formatted (as decimals)}

\item{showPlusSign}{Whether or not to show "+" in front of positive percentages}

\item{digits}{How many digits to round to and display}
}
\value{
Vector of characters
}
\description{
Format values with "\%" at end, x digits after the decimal and comma separators at thousands, optional "+" in front
}
\author{
Stefanie Molin
}
