% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_check_library.R
\name{get_max_dates}
\alias{get_max_dates}
\title{Pull Max Dates Data}
\usage{
get_max_dates(username, password, scope_product, last_try, test_mode,
  root_config_list, regional_config_list, sender, mailbox_password, slack_on)
}
\arguments{
\item{username}{Vertica username}

\item{password}{Vertica password}

\item{scope_product}{Which SCOPE product the code is running in (determines error handling). Choices: "AS", "EXEC", "TS"}

\item{last_try}{Boolean indicating whether or not this is the last try for the script as a whole}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}

\item{slack_on}{Boolean indicating whether or not current SCOPE configuration can post to slack. Defaults to FALSE}
}
\value{
Dataframe of results
}
\description{
Pull data to check if detection can run.
}
\author{
Stefanie Molin
}
