% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05f_input_data_aero.R
\name{input_data_aero}
\alias{input_data_aero}
\title{Modifies input values for aerodynamic parameters within input_data_default.csv}
\usage{
input_data_aero(
  zo = 0.25,
  d = 1.34,
  Cd = 0.3,
  rb = 10,
  CR = 0.35,
  CD1 = 20.6,
  Psicor = 0.2,
  CSSOIL = 0.01,
  rbs = 10,
  rwc = 0
)
}
\arguments{
\item{zo}{roughness length for momentum of the canopy (m)}

\item{d}{displacement height (m)}

\item{Cd}{leaf drag coefficient}

\item{rb}{leaf boundary resistance (s m-1)}

\item{CR}{Verhoef et al. (1997) Drag coefficient for isolated tree}

\item{CD1}{Verhoef et al. (1997) fitting parameter}

\item{Psicor}{Verhoef et al. (1997) Roughness layer correction}

\item{CSSOIL}{Drag coefficient for soil Verhoef et al. (1997) (from Aerodynamic)}

\item{rbs}{soil boundary layer resistance (from Aerodynamic) (s m-1)}

\item{rwc}{within canopy layer resistance}

\item{path}{filepath of the filenames.csv}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for aerodynamic parameters inside input_data_default.csv contains.
}
\examples{
input_data_aero(zo = 0.3)
}
