% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_angles.R
\name{input_data_angles}
\alias{input_data_angles}
\title{Modifies input values for solar angle within input_data_default.csv}
\usage{
input_data_angles(tts = 30, tto = 0, psi = 0)
}
\arguments{
\item{tts}{solar zenith angle (deg)}

\item{tto}{observer zenith angle (deg)}

\item{psi}{azimuthal difference between solar and observation angle; relative azimuth angle (deg)}

\item{path}{filepath of the filenames.csv}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for solar angle parameters inside input_data_default.csv contains.
}
