% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_canopy.R
\name{input_data_canopy}
\alias{input_data_canopy}
\title{Modifies module for canopy properties within input_data_default.csv}
\usage{
input_data_canopy(
  LAI = 3,
  hc = 2,
  LIDFa = -0.35,
  LIDFb = -0.15,
  leafwidth = 0.1,
  Cv = 1,
  crowndiameter = 1
)
}
\arguments{
\item{LAI}{leaf area index (m2 m-2)}

\item{hc}{vegetation height (m)}

\item{LIDFa}{leaf inclination}

\item{LIDFb}{variation in leaf inclination}

\item{leafwidth}{leaf width  (m)}

\item{Cv}{(no description in github)}

\item{crowndiameter}{crown diameter (no units in github)}

\item{path}{filepath of the filenames.csv}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for canopy properties inside input_data_default.csv contains.
}
\examples{
input_data_canopy(LAI = 6.5)
}
