% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_input_data_default.R
\name{input_data_default}
\alias{input_data_default}
\title{General function to modify input_data_default.csv}
\usage{
input_data_default(
  Cab = 40,
  Cca = 10,
  Cdm = 0.012,
  Cw = 0.009,
  Cs = 0,
  Cant = 1,
  Cp = 0,
  Cbc = 0,
  N = 1.5,
  rho_thermal = 0.01,
  tau_thermal = 0.01,
  Vcmax25 = 60,
  BallBerrySlope = 8,
  BallBerry0 = 0.01,
  Type = 0,
  kV = 0.64,
  Rdparam = 0.015,
  Kn0 = 2.48,
  Knalpha = 2.83,
  Knbeta = 0.114,
  Tyear = 15,
  beta = 0.51,
  kNPQs = 0,
  qLs = 1,
  stressfactor = 1,
  fqe = 0.01,
  spectrum = 1,
  rss = 500,
  rs_thermal = 0.06,
  cs = 1180,
  rhos = 1800,
  lambdas = 1.55,
  SMC = 25,
  BSMBrightness = 0.5,
  BSMlat = 25,
  BSMlon = 45,
  LAI = 3,
  hc = 2,
  LIDFa = -0.35,
  LIDFb = -0.15,
  leafwidth = 0.1,
  Cv = 1,
  crowndiameter = 1,
  z = 5,
  Rin = 600,
  Ta = 20,
  Rli = 300,
  p = 970,
  ea = 15,
  u = 2,
  Ca = 410,
  Oa = 209,
  zo = 0.25,
  d = 1.34,
  Cd = 0.3,
  rb = 10,
  CR = 0.35,
  CD1 = 20.6,
  Psicor = 0.2,
  CSSOIL = 0.01,
  rbs = 10,
  rwc = 0,
  startDOY = 20060618,
  endDOY = 20300101,
  LAT = 51.55,
  LON = 5.55,
  timezn = 1,
  tts = 30,
  tto = 0,
  psi = 0
)
}
\arguments{
\item{Cab}{Total Chla + Chlb}

\item{Cca}{Total Carotenoid content}

\item{Cdm}{Leaf dry matter}

\item{Cw}{Leaf water equivalent}

\item{Cs}{Leaf senescent material fraction}

\item{Cant}{Leaf anthocyanin content}

\item{Cp}{Protein}

\item{Cbc}{?}

\item{N}{leaf thickness parameters}

\item{rho_thermal}{broadband thermal reflectance}

\item{tau_thermal}{broadband thermal transmittance}

\item{Vcmax25}{maximum carboxylation efficiency (at optimum temperature of 25 deg C, former Vcmo)}

\item{BallBerrySlope}{slope of Ball-Berry stomatal conductance (former m)}

\item{BallBerry0}{intercept of Ball-Berry stomatal conductance model}

\item{Type}{Photochemical pathway: 0 => C3; 1 => C4}

\item{kV}{Extinction coefficient for Vcmax in the vertical (maximum at the top). 0 for uniform Vcmax}

\item{Rdparam}{Respiration = Rdparam * Vcmax}

\item{Kn0}{Kn0: parameter for empirical Kn (NPQ) model}

\item{Knalpha}{alpha parameter for empirical Kn (NPQ) model}

\item{Knbeta}{beta parameter for empirical Kn (NPQ) model}

\item{Tyear}{mean annual temperature}

\item{beta}{fraction of photons partitioned to PSII}

\item{kNPQs}{rate constant of sustained thermal dissipation}

\item{qLs}{fraction of cuntional reaction centers}

\item{stressfactor}{optional input: stress factor to reduce Vcmax (for example soil moisture, leaf age)}

\item{fqe}{fluorescence quantum yield efficiency at photosynstem level}

\item{spectrum}{spectrum number (column in the database soil_file)}

\item{rss}{soil resistance for evaporation from the pore space}

\item{rs_thermal}{broadband soil reflectance in the thermal range (1 - emissivity)}

\item{cs}{specific heat capacity of the soil}

\item{rhos}{specific mass of the soil}

\item{lambdas}{heat conductivity of the soil}

\item{SMC}{volumetric soil moisture content in the root zone}

\item{BSMBrightness}{BSM model parameter for soil brightness}

\item{BSMlat}{BSM model parameter (lat)}

\item{BSMlon}{BSM model parameter (long)}

\item{LAI}{leaf area idex}

\item{hc}{vegetation height}

\item{LIDFa}{leaf inclination}

\item{LIDFb}{variation in leaf inclination}

\item{leafwidth}{leaf width}

\item{Cv}{?}

\item{crowndiameter}{default to 1}

\item{z}{measurement height of meteorological data (default to 5)}

\item{Rin}{broadband shortwave radiation}

\item{Ta}{air temperature}

\item{Rli}{broadband incoming longwave radiation}

\item{p}{air pressure}

\item{ea}{atmospheric vapour pressure}

\item{u}{wind speed at height z}

\item{Ca}{atmospheric CO2 concentration}

\item{Oa}{atmospheric O2 concentration}

\item{zo}{roughness length for momentum of the canopy}

\item{d}{displacement height}

\item{Cd}{leaf drag coefficient}

\item{rb}{leaf boundary resistance}

\item{CR}{Drag coefficient for isolated tree}

\item{CD1}{fitting paramter}

\item{Psicor}{Roughness layer correction}

\item{CSSOIL}{Drag coefficient for soil}

\item{rbs}{soul boundary layer resistance (from Aerodynamic)}

\item{rwc}{within canopy layer resistance}

\item{startDOY}{Julian day start of simulations}

\item{endDOY}{Julian day end of simulations}

\item{LAT}{latitude}

\item{LON}{longitude}

\item{timezn}{east of Greenwich}

\item{tts}{solar zenith angle}

\item{tto}{observer zenigh angle}

\item{psi}{azimuthal difference between solar and observation angle; relative azimuth angle}
}
\description{
The input_dat_default.csv contains input values for leaf (PROSPECT and biochemistry), canopy, soil, meteorology, aerodynamics, time series and angles. This function helps you modify input values for all modules.
}
