% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_PROSPECT.R
\name{input_data_PROSPECT}
\alias{input_data_PROSPECT}
\title{Modifies PROSPECT module within input_data_default.csv}
\usage{
input_data_PROSPECT(
  Cab = 40,
  Cca = 10,
  Cdm = 0.012,
  Cw = 0.009,
  Cs = 0,
  Cant = 1,
  Cp = 0,
  Cbc = 0,
  N = 1.5,
  rho_thermal = 0.01,
  tau_thermal = 0.01
)
}
\arguments{
\item{Cab}{Chlorophyll a b content (ug cm-2)}

\item{Cca}{Carotenoid content. Usually 25\% of Cab if options.Cca_function_of_Cab (ug cm-2)}

\item{Cdm}{Dry matter content (g cm-2)}

\item{Cw}{leaf water equivalent layer (cm)}

\item{Cs}{senescent material fraction}

\item{Cant}{Anthocyanins (ug cm-2)}

\item{Cp}{protein (ug cm-2)}

\item{Cbc}{?}

\item{N}{leaf thickness parameters}

\item{rho_thermal}{broadband thermal reflectance}

\item{tau_thermal}{broadband thermal transmitance}

\item{path}{filepath of the filenames.csv}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for PROSPECT module inside input_data_default.csv contains.
}
