% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_soil.R
\name{input_data_soil}
\alias{input_data_soil}
\title{Modifies module for soil properties within input_data_default.csv}
\usage{
input_data_soil(
  path = "../SCOPE/input/input_data_default.csv",
  spectrum = 1,
  rss = 500,
  rs_thermal = 0.06,
  cs = 1180,
  rhos = 1800,
  lambdas = 1.55,
  SMC = 25,
  BSMBrightness = 0.5,
  BSMlat = 25,
  BSMlon = 45
)
}
\arguments{
\item{path}{filepath of the filenames.csv}

\item{spectrum}{spectrum number (column in the database soil_file)}

\item{rss}{soil resistance for evaporation from the pore space (s m-1)}

\item{rs_thermal}{broadband soil reflectance in the thermal range \link{1 - emissivity}}

\item{cs}{specific heat capacity of the soil (J kg-1 K-1)}

\item{rhos}{specific mass of the soil (kg m-3)}

\item{lambdas}{heat conductivity of the soil (J m-1 K-1)}

\item{SMC}{volumetric soil moisture content in the root zone}

\item{BSMBrightness}{BSM model parameter for soil brightness}

\item{BSMlat}{BSM model parameter 'lat'}

\item{BSMlon}{BSM model parameter 'long'}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for soil properties inside input_data_default.csv contains.
}
