% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_check_library.R
\name{late_data_notifications}
\alias{late_data_notifications}
\title{Handle Late Data Notifications}
\usage{
late_data_notifications(last_try, scope_product, root_config_list,
  regional_config_list, test_mode, sender, mailbox_password, slack_on = FALSE)
}
\arguments{
\item{last_try}{Boolean whether or not this is the last attempt to run SCOPE}

\item{scope_product}{Which SCOPE product the code is running in (determines recipients). Choices: "AS", "EXEC", "TS"}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}

\item{slack_on}{Boolean indicating whether or not to send notifications to slack}
}
\description{
Handles late data notifications when it is the last try and when it will try again later.
}
\author{
Stefanie Molin
}
