% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_par_constant.R
\name{load_par_constant}
\alias{load_par_constant}
\title{Loads constant parameters for time series input}
\usage{
load_par_constant()
}
\description{
The load_par_constant() function is designed to prompt the user to select a file containing constant parameters the timeseries input table. The function then reads the data from the selected file using the readRDS() function and returns the data as a variable.
}
