% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_load_refl_data.r
\name{load_refl_data}
\alias{load_refl_data}
\title{Load Reflectance Data}
\usage{
load_refl_data(folder = ".", start_time, end_time)
}
\arguments{
\item{folder}{Character string specifying the folder
containing the data files (default is current directory).}

\item{start_time}{Character string specifying the
starting time in the format YYYY-DOY.}

\item{end_time}{Character string specifying the
ending time in the format YYYY-DOY.}
}
\value{
A data frame containing the combined reflectance data.
}
\description{
This function loads reflectance data from a file,
combines it with year and DoY from aPAR.csv,
and returns a data frame.
}
\examples{
data <- load_refl_data(".", "2022-120", "2022-130")
}
