% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{log_to_history}
\alias{log_to_history}
\title{Log to the History File}
\usage{
log_to_history(history_file, script_name, geo = "")
}
\arguments{
\item{history_file}{Path of the YAML file containing the run history}

\item{script_name}{Name of the script currently running}

\item{geo}{Location currently running}
}
\value{
None, updates the history file if necessary
}
\description{
Log when and which scripts have run
}
\author{
Francois Pillot

Stefanie Molin
}
