% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vertica.R
\name{modify_query}
\alias{modify_query}
\title{Modify Query}
\usage{
modify_query(query, sub_list)
}
\arguments{
\item{query}{The query string.}

\item{sub_list}{A list where the names are the words to replace in the query and the values is what to change them to.}
}
\value{
String containing the modified query.
}
\description{
Modify a query by replacing keywords with new values
}
\author{
Stefanie Molin
}
