% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_explain.R
\name{omni_explain}
\alias{omni_explain}
\title{Explain Feature Importance}
\usage{
omni_explain(omni_data, phenotype)
}
\arguments{
\item{omni_data}{Integrated omics data (OmniScope object)}

\item{phenotype}{Numeric or factor vector}
}
\value{
A data frame of feature importance scores
}
\description{
Estimates basic feature importance using variance or correlation with phenotype.
}
