% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni_visualize.R
\name{omni_visualize}
\alias{omni_visualize}
\title{Visualize the Causal Graph}
\usage{
omni_visualize(causal_graph, max_nodes = 50)
}
\arguments{
\item{causal_graph}{Output from omni_causal()}

\item{max_nodes}{Maximum number of nodes to plot (default = 50)}
}
\description{
Creates a simple interactive-like plot of causal relationships among features.
}
