% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{percent_difference}
\alias{percent_difference}
\title{Percent Difference}
\usage{
percent_difference(before_values, after_values, format = TRUE, ...)
}
\arguments{
\item{before_values}{Vector of values in "before" condition}

\item{after_values}{Vector of values in "after" condition}

\item{format}{Whether or not to format as a percent}

\item{...}{Additional arguments to be passed to format_percent()}
}
\value{
Vector of characters
}
\description{
Calculate percent difference and return it formatted if desired
}
\author{
Stefanie Molin
}
