% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_library.R
\name{prep_slack}
\alias{prep_slack}
\title{Prep for Slack Notifications}
\usage{
prep_slack(root_config_list)
}
\arguments{
\item{root_config_list}{List from the root_config file.}
}
\description{
Sets up the slack bot
}
\author{
Stefanie Molin
}
