% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection_library.R
\name{pull_AS_data}
\alias{pull_AS_data}
\title{Pull AS Data}
\usage{
pull_AS_data(username, password, last_try, test_mode, root_config_list,
  regional_config_list, sender, mailbox_password, slack_on = FALSE)
}
\arguments{
\item{username}{Vertica username}

\item{password}{Vertica password}

\item{last_try}{Boolean indicating whether or not this is the last try for the script as a whole}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}

\item{slack_on}{Boolean indicating whether or not current SCOPE configuration can post to slack. Defaults to FALSE}
}
\value{
Dataframe of results
}
\description{
Pull data in the proper format for AS anomaly detection
}
\author{
Stefanie Molin
}
