% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis_data_collection.R
\name{pull_AS_metrics_for_metis}
\alias{pull_AS_metrics_for_metis}
\title{Pull AS KPI Evolution Data for Metis}
\usage{
pull_AS_metrics_for_metis(username, password, dest_path,
  country_conditions = "em.country IN ('US', 'CA') AND em.cost_center_country NOT IN ('LM') AND em.cost_center_activity IN ('TIE', 'IAP')",
  tiering = "ranking IN ('TIER 1')",
  job_name = "em.job_name IN ('T1 ACCOUNT STRATEGIST - FARMING')")
}
\arguments{
\item{username}{Vertica username}

\item{password}{Vertica password}

\item{dest_path}{Path to save the CSV to}

\item{country_conditions}{The country conditions to use in the query (see the .yml files)}

\item{tiering}{The client ranking to use in the query (see the .yml files)}

\item{job_name}{The job name conditions to use in the query (see the .yml files)}
}
\value{
A CSV saved in the dest_path
}
\description{
Queries Vertica for same data SCOPE uses to check for anomalies and 
outputs it ready for Metis's database.
}
\note{
Defaults will query US data
}
\author{
Stefanie Molin
}
