% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection_library.R
\name{pull_Exec_data}
\alias{pull_Exec_data}
\title{Pull Exec Data}
\usage{
pull_Exec_data(territory_level, username, password, last_try, test_mode,
  root_config_list, regional_config_list, sender, mailbox_password)
}
\arguments{
\item{territory_level}{Boolean TRUE will pull territory-level stats, FALSE pulls client-level}

\item{username}{Vertica username}

\item{password}{Vertica password}

\item{last_try}{Boolean indicating whether or not this is the last try for the script as a whole}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}
}
\value{
Dataframe of results
}
\description{
Pull data in the proper format for Exec anomaly detection
}
\author{
Stefanie Molin
}
