% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis_data_collection.R
\name{pull_exec_metrics_for_metis}
\alias{pull_exec_metrics_for_metis}
\title{Pull Exec RexT Evolution Data for Metis}
\usage{
pull_exec_metrics_for_metis(username, password, dest_path)
}
\arguments{
\item{username}{Vertica username}

\item{password}{Vertica password}

\item{dest_path}{Path to save the CSV to}
}
\value{
A CSV saved in the dest_path
}
\description{
Queries Vertica for same data SCOPE uses to check for anomalies and 
outputs it ready for Metis's database.
}
\note{
Defaults will query Americas data
}
\author{
Stefanie Molin
}
