% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vertica.R
\name{query_try_block}
\alias{query_try_block}
\title{Query Try Block}
\usage{
query_try_block(query, df_name, username, password, timeout, scope_product,
  last_try, test_mode, root_config_list, regional_config_list, sender,
  mailbox_password, slack_on = FALSE, send_error_email = TRUE)
}
\arguments{
\item{query}{Vertica SQL query as a string}

\item{df_name}{Name of the dataframe you are querying for (this will only be used for logging purposes)}

\item{username}{Vertica username}

\item{password}{Vertica password}

\item{timeout}{In seconds, how long should R keep trying to query Vertica for?}

\item{scope_product}{Which SCOPE product the code is running in (determines error handling). Choices: "AS", "EXEC", "TS"}

\item{last_try}{Boolean indicating whether or not this is the last try for the script as a whole}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}

\item{slack_on}{Boolean indicating whether or not current SCOPE configuration can post to slack. Defaults to FALSE}

\item{send_error_email}{Boolean indicating whether or not an error email should be sent out if the query fails.
Defaults to TRUE, set to FALSE if the results of the query aren't required to finish running successfully.}
}
\value{
FALSE if the query fails and no error email will be sent, otherwise dataframe of results
}
\description{
Try to query Vertica for a specific query and handle notifications if not working
}
\author{
Stefanie Molin
}
