% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vertica.R
\name{query_vertica}
\alias{query_vertica}
\title{Query Vertica}
\usage{
query_vertica(username, query, password)
}
\arguments{
\item{username}{Vertica username}

\item{query}{Vertica SQL query as a string}

\item{password}{Vertica password}
}
\value{
Dataframe with the results
}
\description{
Query Vertica and return a dataframe
}
\author{
Stefanie Molin
}
