% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vertica.R
\name{read_query_from_file}
\alias{read_query_from_file}
\title{Read Query from a File}
\usage{
read_query_from_file(filepath)
}
\arguments{
\item{filepath}{Path to the query (as a string)}
}
\value{
String containing the query.
}
\description{
Read query from a textfile and collapse lines into 1 string
}
\author{
Stefanie Molin
}
