% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomaly.R
\name{record_anomaly}
\alias{record_anomaly}
\title{Record Anomaly}
\usage{
record_anomaly(current_alerts, alert_to_add)
}
\arguments{
\item{current_alerts}{String containing the alerts already flagged}

\item{alert_to_add}{String containing the newest flagged alert.}
}
\value{
String
}
\description{
Record the anomaly adding to a string of anomalies
}
\author{
Stefanie Molin
}
