% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCOPE.R
\name{run_scope}
\alias{run_scope}
\title{Runs SCOPE.m in R}
\usage{
run_scope(path)
}
\arguments{
\item{path}{is the location of the directory where SCOPE.m is stored.}
}
\value{
The default value is "../../SCOPE/SCOPE.m". If this does not work, this means either you modify the name of your SCOPE model (\emph{which is not recommended}) or your current working directory is somewhere else in the universe. In any case, just copy and paste the original filepath of SCOPE.m enclosed in quotations in case the default value does not work.
}
\description{
This function runs SCOPE.m in R language. Copy and paste the location of the file SCOPE.m enclosed in open-close quotations.
}
