% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scope_output.R
\name{scope_output}
\alias{scope_output}
\title{Lists outputs from SCOPE simulations}
\usage{
scope_output(output_dir = "../.././SCOPE/output/", except, subset)
}
\arguments{
\item{output_dir}{filepath of SCOPE output directory}

\item{except}{files to exclude (works with regex too)}

\item{subset}{files to subset}
}
\description{
Lists outputs from SCOPE simulations
}
