% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoper.R
\docType{package}
\name{scoper}
\alias{scoper}
\alias{scoper-package}
\title{The SCOPer package}
\description{
Provides a computational framework for B cell clones identification
from adaptive immune receptor repertoire sequencing (AIRR-Seq) datasets. 
Three models are included (identical, hierarchical, and spectral) 
which perform clustering among sequences of B cell receptors 
(BCRs, also referred to as Immunoglobulins, (Igs)) that 
share the same V gene, J gene and junction length.
}
\section{SCOPer}{


\itemize{
  \item  \link{defineClonesScoper}:  Clustering sequences into clonal groups.
}
}

\references{
\enumerate{
  \item  Nouri N and Kleinstein SH (2018). A spectral clustering-based method for identifying clones
  from high-throughput B cell repertoire sequencing data. Bioinformatics, 34(13):i341-i349.
  \item Gupta NT, et al. (2017). Hierarchical clustering can identify B cell clones with high 
  confidence in Ig repertoire sequencing data. The Journal of Immunology, 198(6):2489-2499.
}
}
