% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_as_summary_stats}
\alias{send_as_summary_stats}
\title{Send AS Summary Stats}
\usage{
send_as_summary_stats(data, test_mode, root_config_list, regional_config_list,
  sender, mailbox_password, slack_on = FALSE)
}
\arguments{
\item{data}{Dataframe of accounts triggered and for which metric}

\item{test_mode}{Boolean whether or not SCOPE is running in test mode or not}

\item{root_config_list}{List from the root_config file.}

\item{regional_config_list}{List from the regional_config file.}

\item{sender}{Email address to send from}

\item{mailbox_password}{Password for the sender}

\item{slack_on}{Boolean indicating whether or not information should be sent to slack}
}
\description{
Send summary of AS alerts via email (and slack if desired).
}
\author{
Stefanie Molin
}
