% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_email}
\alias{send_email}
\title{Send Email}
\usage{
send_email(from, to, cc = NULL, bcc = NULL, max_retry_time = 300,
  password, timeout = 3e+05, attachments = NULL, subject = "",
  body = "", encoding = "utf-8", html = FALSE, inline = FALSE,
  send = TRUE)
}
\arguments{
\item{from}{Sender address}

\item{to}{Email recipient address}

\item{cc}{CC recipient address}

\item{bcc}{BCC recipient address}

\item{max_retry_time}{In seconds, how long to retry sending email}

\item{password}{Password for the from address}

\item{timeout}{In seconds, how long to wait in case of email server timeout}

\item{attachments}{Path(s) to files to attach.}

\item{subject}{The subject of the email}

\item{body}{The body of the email}

\item{encoding}{The encoding to use in the email}

\item{html}{Boolean indicating whether or not the email contents will be in HTML}

\item{inline}{Boolean indicating whether or not the email contains inline images.}

\item{send}{Boolean indicating whether or not to send the email.}
}
\value{
Boolean indicating whether or not sending the email was successful
}
\description{
Wrapper for the send.mail function, adding a loop for multiple trials
}
\author{
Francois Pillot
}
