% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_email_once}
\alias{send_email_once}
\title{Send Email Once}
\usage{
send_email_once(from, to, cc = NULL, bcc = NULL, max_retry_time = 300,
  password, timeout = 3e+05, attachments = NULL, subject = "",
  body = "", encoding = "utf-8", html = FALSE, inline = FALSE,
  max_retry_message = "", history_file)
}
\arguments{
\item{from}{Sender address}

\item{to}{Email recipient address}

\item{cc}{CC recipient address}

\item{bcc}{BCC recipient address}

\item{max_retry_time}{In seconds, how long to retry sending email}

\item{password}{Password for the from address}

\item{timeout}{In seconds, how long to wait in case of email server timeout}

\item{attachments}{Path(s) to files to attach.}

\item{subject}{The subject of the email}

\item{body}{The body of the email}

\item{encoding}{The encoding to use in the email}

\item{html}{Boolean indicating whether or not the email contents will be in HTML}

\item{inline}{Boolean indicating whether or not the email contains inline images.}

\item{max_retry_message}{What to print to logs if function has exhausted its retries}

\item{history_file}{Path to the YAML history file containing the history of sent emails}
}
\value{
Sends email if not sent before and logs it, otherwise doesn't
}
\description{
Ensure that each recpient gets one email max
}
\author{
Francois Pillot
}
