% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_no_alerts_found_email}
\alias{send_no_alerts_found_email}
\title{Send No Alerts Found Email}
\usage{
send_no_alerts_found_email(scope_product, test_mode, root_config_list,
  regional_config_list, sender, mailbox_password)
}
\arguments{
\item{scope_product}{Which SCOPE product the code is running in (determines recipients). Choices: "AS", "EXEC", "TS"}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{root_config_list}{List read from the root config YAML file}

\item{regional_config_list}{List read from the regional config YAML file}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}
}
\value{
Sends late data email if not sent before and logs it, otherwise doesn't
}
\description{
Send email to maintainers that there were no alerts triggered
}
\author{
Stefanie Molin
}
