% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_technical_error_email}
\alias{send_technical_error_email}
\title{Send Technical Error Email}
\usage{
send_technical_error_email(scope_product, root_config, regional_config,
  test_mode, sender, mailbox_password)
}
\arguments{
\item{scope_product}{Which SCOPE product the code is running in (determines recipients). Choices: "AS", "EXEC", "TS"}

\item{root_config}{List read from the root config YAML file}

\item{regional_config}{List read from the regional config YAML file}

\item{test_mode}{Boolean indicating whether or not SCOPE is in test mode}

\item{sender}{Address of the sender of the error emails.}

\item{mailbox_password}{Password of the sender}
}
\value{
Sends email if not sent before and logs it, otherwise doesn't
}
\description{
Send the technical error email to the appropriate recipients based on which SCOPE is running
}
\author{
Stefanie Molin
}
