% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_library.R
\name{send_ts_summary_stats}
\alias{send_ts_summary_stats}
\title{Send TS Summary Stats}
\usage{
send_ts_summary_stats(eventSiteAlertsDF, test_mode, root_config_list,
  regional_config_list, sender, mailbox_password, slack_on = FALSE)
}
\arguments{
\item{eventSiteAlertsDF}{Dataframe of tag-level triggered alerts}

\item{test_mode}{Boolean whether or not SCOPE is running in test mode or not}

\item{root_config_list}{List from the root_config file.}

\item{regional_config_list}{List from the regional_config file.}

\item{sender}{Email address to send from}

\item{mailbox_password}{Password for the sender}

\item{slack_on}{Boolean whether or not to send alerts to slack channel}
}
\description{
Send summary of TS alerts via email.
}
\author{
Stefanie Molin
}
