% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_library.R
\name{send_ts_summary_to_slack}
\alias{send_ts_summary_to_slack}
\title{Send TS Alerts Summary to Slack}
\usage{
send_ts_summary_to_slack(accounts_triggered, regional_config_list,
  root_config_list, ranking, test_mode)
}
\arguments{
\item{accounts_triggered}{Dataframe of accounts triggered and for which metric}

\item{regional_config_list}{List from the regional_config file.}

\item{root_config_list}{List from the root_config file.}

\item{ranking}{Ranking of the account as found in Vertica (all caps)}

\item{test_mode}{Boolean whether or not SCOPE is running in test mode or not}
}
\description{
Send summary of TS alerts to slack.
}
\author{
Stefanie Molin
}
