% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_set_filenames.R
\name{set_filenames}
\alias{set_filenames}
\title{Modify filenames.csv}
\usage{
set_filenames(
  simulation_name,
  dataset_dir,
  meteo_ec_csv,
  t = TRUE,
  Rin = TRUE,
  Rli = FALSE,
  p = FALSE,
  Ta = TRUE,
  u = FALSE,
  ea = FALSE,
  RH = TRUE,
  VPD = FALSE,
  tts = TRUE,
  tto = TRUE,
  psi = TRUE,
  Cab = TRUE,
  Cca = TRUE,
  Cdm = TRUE,
  Cw = TRUE,
  Cs = TRUE,
  Cant = TRUE,
  SMC = FALSE,
  BSMBrightness = FALSE,
  BSMlat = FALSE,
  BSMlon = FALSE,
  LAI = TRUE,
  hc = FALSE,
  LIDFa = TRUE,
  LIDFb = TRUE,
  z = FALSE,
  Ca = FALSE,
  Vcmax25 = TRUE,
  BallBerrySlope = TRUE
)
}
\arguments{
\item{simulation_name}{Lets you change the simulation name.}

\item{dataset_dir}{Lets you specify the directory of the
dataset containing the input files for time series simulations}

\item{meteo_ec_csv}{Lets you specify the name of the .csv file
that contains the meteorological information to be used in time
series runs (more of this in the create_ts_input() function)}

\item{t}{Julian day (decimal)
zenith angle if th value is TRUE. Place a FALSE value if
latitude and longitude are specified in the input.filename.csv}

\item{Rin}{broadband incoming shortwave radiation (0.4-2.5 um) (W m-2)}

\item{Rli}{broadband incoming longwave radiation (2.5-50 um) (W m-2)}

\item{p}{air pressure (hPa)}

\item{Ta}{Air temperature}

\item{u}{wind speed at height z (m s-1)}

\item{ea}{atmospheric vapour pressure (hPa)}

\item{RH}{relative humidity}

\item{VPD}{vapor pressure deficit}

\item{tts}{Solar zenith angle}

\item{tto}{Observer zenith angle}

\item{psi}{Relative azimuth angle: the azimuthal difference
between solar and observation angle.}

\item{Cab}{Chlorophyll a b content (ug cm-2)}

\item{Cdm}{dry matter content}

\item{Cw}{leaf water equivalent layer (cm)}

\item{Cs}{senescent material fraction}

\item{Cant}{Anthocyanins (ug cm-2)}

\item{SMC}{volumetric soil moisture content in the root zone}

\item{BSMBrightness}{BSM model parameter for soil brightness}

\item{BSMlat}{BSM model parameter ‘lat’}

\item{BSMlon}{BSM model parameter ‘long’}

\item{LAI}{leaf area index (m-2 m-2)}

\item{hc}{vegetation height (m)}

\item{LIDFa}{leaf inclination}

\item{LIDFb}{variation in leaf inclination}

\item{z}{measurement height of meteorological data}

\item{Ca}{atmospheric CO2 concentration (ppm)}

\item{Vcmax25}{maximum carboxylation capacity (at optimum
temperature of 25C, former Vcmo) (umol m-2 s-1)}

\item{BallBerrySlope}{slope of Ball-Berry stomatal
conductance model (former m)}

\item{Car}{Carotenoid content. Usually 25\% of Cab if
options.Cca_function_of_Cab (ug cm-2)}
}
\description{
This function helps you modify filenames.csv
}
