% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_options}
\alias{set_options}
\title{Modify setoptions.csv}
\usage{
set_options(verify = TRUE, simulation = 0)
}
\arguments{
\item{verify}{Lets you switch verification mode.}

\item{simulation}{Defines rules for input reading.}
}
\value{
\strong{MAIN}

\strong{verify} Default value is TRUE which means the run will
compare output to the verification dataset to test the integrity of
your SCOPE copy. FALSE will switch off the verification mode.
Note: You would need to switch off the verification mode if you
want to do actual simulations.

\strong{simulation} Default value is 0 which is equal to
individual runs. Other options can be 1 (time series runs) or 2
(for look-up table).
}
\description{
This function helps you modify setoptions.csv. It reads in a CSV
file of options for SCOPE, modifies the values of those options
based on the input parameters, and writes the updated options back
to the CSV file. The function takes several parameters with default
values, primarily \code{verify} and \code{simulation}, which control aspects
of the SCOPE simulation. This function is intended to make it easy
to customize the SCOPE simulation options without modifying the
underlying MATLAB code directly.
}
\details{
For User's Comfort, run this code with arguments verify = FALSE
}
