% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paramFile.R
\name{set_paramFile}
\alias{set_paramFile}
\title{Modify set_parameter_filenames.csv}
\usage{
set_paramFile(input = "input_data_latin_hypercube.csv")
}
\arguments{
\item{input}{Change the name of the .csv file where input data will be grabbed by SCOPE.}

\item{path}{Filepath of the directory where set_parameter_filenames.csv is stored}
}
\value{
\strong{path} Value is a string that represents filepath of the directory where set_parameter_filenames.csv is stored. Default value is "../../SCOPE/set_parameter_filenames.csv" which should work even if you don't specify the filepath. In case it does not, copy and paste the long filepath as a value of this argument.

\strong{input} By default, the value of this argument is "input_data_latin_hypercube.csv" which is also the default setting in SCOPE. Change this to "input_data_default.csv" if you want to simulate individual runs or time series experiments. Note: You can also changed the default values inside the input_data_default.csv using the function input_data_default().
}
\description{
This function helps you modify set_parameter_filenames.csv input file.
}
