% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{site_events_7D}
\alias{site_events_7D}
\title{Site Events Last 7 Days Dataframe}
\usage{
site_events_7D(data, timezone)
}
\arguments{
\item{data}{The dataframe containing 2 columns: date and site events (naming doesn't matter but order does)}

\item{timezone}{The timezone to create the additional dates in (if necessary)}
}
\value{
Dataframe of site events with first column "day" and second column "total_events"
}
\description{
Return in 7 day dataframe, filling dates with "NA" if not available
}
\author{
Stefanie Molin
}
