% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_library.R
\name{slack_no_alerts_found}
\alias{slack_no_alerts_found}
\title{Slack Notifications for No Alerts}
\usage{
slack_no_alerts_found(test_mode, root_config_list, regional_config_list,
  scope_product)
}
\arguments{
\item{test_mode}{Boolean whether or not SCOPE is running in test mode or not}

\item{root_config_list}{List from the root_config file.}

\item{regional_config_list}{List from the regional_config file.}

\item{scope_product}{Which SCOPE product this is for i.e. AS, TS, etc.}
}
\description{
Posts to slack channel(s) saying there are no alerts
}
\author{
Stefanie Molin
}
