% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_sourceEntireFolder.R
\name{sourceEntireFolder}
\alias{sourceEntireFolder}
\title{finds all .R files within a folder and soruces them}
\usage{
sourceEntireFolder(folderName, verbose = FALSE, showWarnings = TRUE)
}
\arguments{
\item{folderName}{For looping the individual R script, this folder should contain either species or cluster .R to run individual case}

\item{verbose}{shows the sourcing}

\item{showWarnings}{show warnings whether there are R files in the folder}
}
\description{
finds all .R files within a folder and soruces them
}
