% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{thousands_formatter}
\alias{thousands_formatter}
\title{Thousands Formatter}
\usage{
thousands_formatter(x)
}
\value{
Formatted number(s)
}
\description{
Format numbers with thousands place to say "K" at end and "M" for millions
}
\author{
Stefanie Molin
}
