% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{transparent_graph_theme}
\alias{transparent_graph_theme}
\title{Transparent Graph Theme}
\usage{
transparent_graph_theme()
}
\value{
A ggplot2 theme.
}
\description{
Make a ggplot2 graph theme for SCOPE emails
}
\author{
Stefanie Molin
}
