% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_graphs.R
\name{ts_alert_graphs}
\alias{ts_alert_graphs}
\title{TS Alert Graphs}
\usage{
ts_alert_graphs(tagAlerts, data, folder_name)
}
\arguments{
\item{tagAlerts}{Dataframe containing 1 row per alert with the client name (or client tag combo) and the alerts column.}

\item{data}{The dataframe containing the daily data used to run the model and check for alerts.}

\item{folder_name}{Folder name of where to save the plots}
}
\value{
Graphs for all alerts using the given model.
}
\description{
Generate graphs for all alerts using a given model for visual inspection.
}
\author{
Stefanie Molin
}
