% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{ts_alerts_graph_list}
\alias{ts_alerts_graph_list}
\title{TS Alerts Graph List}
\usage{
ts_alerts_graph_list(graph_stats)
}
\arguments{
\item{graph_stats}{Dataframe of data for a particular TS}
}
\value{
A list containing the names of the saved ggplots indexed by TS name
}
\description{
Make graphs for all TS alerts and return a list with their names indexed by TS
}
\author{
Stefanie Molin
}
