% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{ts_graph_and_save}
\alias{ts_graph_and_save}
\title{TS Graph and Save}
\usage{
ts_graph_and_save(df, plotname, FUN, ...)
}
\arguments{
\item{df}{Dataframe containing the site events data for the advertiser.}

\item{plotname}{String for the name you want to save the plot as; this will be returned upon successful completion}

\item{FUN}{TS graphing function to use; either ts_tag_level_graph or ts_site_level_graph}

\item{...}{Parameters passed to FUN}
}
\value{
The name of the plot after it has been successfully saved.
}
\description{
Make a ggplot2 graph and save, type of graph depends on FUN argument
}
\author{
Stefanie Molin
}
