% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_library.R
\name{TS_transparent_graph}
\alias{TS_transparent_graph}
\title{TS Transparent Graph}
\usage{
TS_transparent_graph(data, combo, encode = FALSE, max_possible_days = 60)
}
\arguments{
\item{data}{Dataframe containing daily data.}

\item{encode}{Whether or not to base64 encode the graph, defaults to FALSE. 
(Needed if sending as part of HTML attachment rather than email body inline images.)}

\item{max_possible_days}{Controls the shifting of the label by creating a ratio of max days allowed in the graph (defaults to 60) and the total days being graphed. 
This makes sure that graphs with less than the max days will shift the label less to keep it on the line.}
}
\value{
A base64 encoded site events plot, or just the plot object (depends on the value of encode).
}
\description{
Make a ggplot2 graph for given metric in TS alerts
}
\author{
Stefanie Molin
}
