% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_library.R
\name{ts_yaml_builder}
\alias{ts_yaml_builder}
\title{TS YAML Builder}
\usage{
ts_yaml_builder(tag_level_alerts, tag_level_stats)
}
\arguments{
\item{tag_level_alerts}{Dataframe containing the alerts for tag-level}

\item{tag_level_stats}{Dataframe of tag-level events}
}
\value{
A list that will transform into YAML when yaml::as.yaml() is used.
}
\description{
Make the graphs and the YAML for the TS alerts
}
\author{
Stefanie Molin
}
