% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_library.R
\name{vector_shrinker}
\alias{vector_shrinker}
\title{Reduce Vector Size}
\usage{
vector_shrinker(vector, last_x_entries)
}
\arguments{
\item{vector}{Data as a vector}

\item{last_x_entries}{Number of entries from the end of the vector desired}
}
\value{
Vector reduced to the last x entries
}
\description{
Reduce a vector to the last x entries
}
\author{
Stefanie Molin
}
