% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_output.R
\name{call_output}
\alias{call_output}
\title{This calls out the filepath of the type of SCOPE output you need for further analysis}
\usage{
call_output(path, type)
}
\arguments{
\item{path}{file path of the directory where output is located.}

\item{type}{type of output you want to analyse. See SCOPE description in github for details}
}
\value{
The default file path assumes that the output from SCOPE is successfully transferred in bcipanama repository.

List of output are as follows: aPAR, apparent_reflectance, Eout_spectrum, Esky, Esun, fluorescence_hemis, fluorescence_ReabsCorr, fluorescence_scalars, fluorescence, fluxes, Lo_spectrum_inclF, Lo_spectrum, Parameters (folder not a .csv), pars_and_input_short, radiation, rdd, rdo, reflectance, resistances, rsd, rso, sigmaF, vegetation, wlF, and wlS
}
\description{
This calls out the filepath of the type of SCOPE output you need for further analysis
}
\examples{
call_output(path = "../bcipanama/7_scope/scope_output", type=fluxes)
}
