% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_library.R
\name{format_datetime}
\alias{format_datetime}
\title{Format Current Date/Time}
\usage{
format_datetime(format_string = "\%Y-\%m-\%d", timezone,
  showTimezone = FALSE)
}
\arguments{
\item{format_string}{The string representing how to format the date. Defaults to YYYY-MM-DD}

\item{timezone}{String for the name of the timezone, i.e. "America/New_York"}

\item{showTimezone}{Boolean indicating whether or not to show the timezone name. Defaults to FALSE (not shown).}
}
\value{
String
}
\description{
Return a string representing the datetime in the specified format for the designated timezone. Optionally show the timezone name.
}
\author{
Stefanie Molin
}
