% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_default.R
\name{input_data_default}
\alias{input_data_default}
\title{General function to modify input_data_default.csv}
\usage{
input_data_default(
  Cab = 40,
  Cca = 10,
  Cdm = 0.012,
  Cw = 0.009,
  Cs = 0,
  Cant = 1,
  Cp = 0,
  Cbc = 0,
  N = 1.5,
  rho_thermal = 0.01,
  tau_thermal = 0.01,
  Vcmax25 = 60,
  BallBerrySlope = 8,
  BallBerry0 = 0.01,
  Type = 0,
  kV = 0.64,
  Rdparam = 0.015,
  Kn0 = 2.48,
  Knalpha = 2.83,
  Knbeta = 0.114,
  Tyear = 15,
  kNPQs = 0,
  qLs = 1,
  stressfactor = 1,
  fqe = 0.01,
  spectrum = 1,
  rss = 500,
  rs_thermal = 0.06,
  cs = 1180,
  rhos = 1800,
  lambdas = 1.55,
  SMC = 25,
  BSMBrightness = 0.5,
  BSMlat = 25,
  BSMlon = 45,
  LAI = 3,
  hc = 2,
  LIDFa = -0.35,
  LIDFb = -0.15,
  leafwidth = 0.1,
  Cv = 1,
  crowndiameter = 1,
  z = 5,
  Rin = 600,
  Ta = 20,
  Rli = 300,
  ea = 15,
  u = 2,
  Ca = 410,
  Oa = 209,
  zo = 0.25,
  d = 1.34,
  Cd = 0.3,
  rb = 10,
  CR = 0.35,
  CD1 = 20.6,
  Psicor = 0.2,
  CSSOIL = 0.01,
  rbs = 10,
  rwc = 0,
  startDOY = 20060618,
  endDOY = 20300101,
  LAT = 51.55,
  LON = 5.55,
  timezn = 1,
  tts = 30,
  tto = 0,
  psi = 0
)
}
\description{
The input_dat_default.csv contains input values for leaf (PROSPECT and biochemistry), canopy, soil, meteorology, aerodynamics, time series and angles. This function helps you modify input values for all modules.
}
