% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoper.R
\docType{package}
\name{scoper-package}
\alias{scoper-package}
\title{scoper: Spectral Clustering-Based Method for Identifying B Cell Clones}
\description{
Provides a computational framework for identification of B cell clones from Adaptive Immune Receptor Repertoire sequencing (AIRR-Seq) data. Three main functions are included (identicalClones, hierarchicalClones, and spectralClones) that perform clustering among sequences of BCRs/IGs (B cell receptors/immunoglobulins) which share the same V gene, J gene and junction length. Nouri N and Kleinstein SH (2018) \doi{ 10.1093/bioinformatics/bty235}. Nouri N and Kleinstein SH (2019) \doi{ 10.1101/788620}. Gupta NT, et al. (2017) \doi{ 10.4049/jimmunol.1601850}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scoper.readthedocs.io}
  \item Report bugs at \url{https://github.com/immcantation/scoper/issues}
}

}
\author{
\strong{Maintainer}: Susanna Marquez \email{susanna.marquez@yale.edu} [contributor]

Authors:
\itemize{
  \item Nima Nouri \email{nima.nouri@yale.edu}
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com}
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Edel Aron \email{edel.aron@yale.edu} [contributor]
  \item Gisela Gabernet \email{gisela.gabernet@yale.edu} [contributor]
  \item Cole Jensen \email{cole.jensen@yale.edu} [contributor]
  \item Huimin Lyu \email{huimin.lyu@yale.edu} [contributor]
}

}
\keyword{internal}
