% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoper.R
\docType{package}
\name{scoper}
\alias{scoper}
\alias{scoper-package}
\title{The SCOPer package}
\description{
Provides a computational framework for unsupervised identification B cell
clones from adaptive immune receptor repertoire sequencing (AIRR-Seq) datasets.
This method is based on spectral clustering of the junction sequences of B cell
receptors (BCRs, Immunoglobulins) that share the same V gene, J gene and
junction length.
}
\section{Spectral Clustering for clOne Partitioning (SCOPer)}{


\itemize{
  \item  \link{defineClonesScoper}:  Clustering sequences into clonal groups.
  \item  \link{analyzeClones}:       Summary statistics and visualization of the
                                     clonal clustering results.
}
}

\references{
\enumerate{
  \item  Nouri N and Kleinstein SH (2018). A spectral clustering-based method for identifying clones
  from high-throughput B cell repertoire sequencing data. Bioinformatics, 34(13):i341-i349.
}
}
